requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Probe interface library
probeinterface>=0.2.0

# 3D file handling
trimesh>=3.10.0
numpy-stl>=2.16.0

# Data validation
jsonschema>=4.0.0
pydantic>=2.0.0

# Visualization (optional)
matplotlib>=3.4.0
plotly>=5.0.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development tools
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Logging and configuration
pyyaml>=6.0
colorlog>=6.6.0

# File handling
h5py>=3.0.0
